refactor(workspace): split workspace shell facade
